refactor signals and tasks
