+++ firstvoices/backend/management/commands/rebuild_index.py
@@ -1,9 +1,9 @@
 from django.core.management.base import BaseCommand
 
 from backend.management.commands._helper import get_valid_index_name, rebuild_index
-from backend.search.indices.dictionary_entry_document import DictionaryEntryDocument
-from backend.search.indices.song_document import SongDocument
-from backend.search.indices.story_document import StoryDocument
+from backend.search.documents.dictionary_entry_document import DictionaryEntryDocument
+from backend.search.documents.song_document import SongDocument
+from backend.search.documents.story_document import StoryDocument
 from backend.search.utils.constants import (
     ELASTICSEARCH_DICTIONARY_ENTRY_INDEX,
     ELASTICSEARCH_SONG_INDEX,
@@ -42,7 +42,7 @@ class Command(BaseCommand):
             rebuild_index(index_name, index_document)
         else:
             self.stdout.write(
-                "Invalid or no index name provided. Building all indices."
+                "Invalid or no index name provided. Building all documents."
             )
             for mapping in self.index_mappings.values():
                 index_name = mapping["index_name"]
