+++ firstvoices/backend/search/__init__.py
@@ -1,1 +1,4 @@
-from .utils.site_signals import *  # noqa F401, F403
+from .signals.dictionary_entry_signals import *  # noqa F401, F403
+from .signals.site_signals import *  # noqa F401, F403
+from .signals.song_signals import *  # noqa F401, F403
+from .signals.story_signals import *  # noqa F401, F403
