+++ firstvoices/backend/search/tasks/site_tasks.py
@@ -1,13 +1,14 @@
 import logging
 
 from celery import shared_task
-from django.core.exceptions import ObjectDoesNotExist
-from django.db.models.signals import post_delete, pre_save
-from django.dispatch import receiver
 from elasticsearch.exceptions import ConnectionError, NotFoundError
 
 from backend.models.sites import Site
-from backend.search.indices import DictionaryEntryDocument, SongDocument, StoryDocument
+from backend.search.documents import (
+    DictionaryEntryDocument,
+    SongDocument,
+    StoryDocument,
+)
 from backend.search.utils.constants import (
     ELASTICSEARCH_DICTIONARY_ENTRY_INDEX,
     ELASTICSEARCH_SONG_INDEX,
@@ -17,27 +18,7 @@ from backend.search.utils.constants import (
     SearchIndexEntryTypes,
 )
 from backend.search.utils.object_utils import get_object_from_index
-from firstvoices.celery import link_error_handler
 from firstvoices.settings import ELASTICSEARCH_LOGGER
-
-
-# If a site's visibility is changed, update all docs from index related to site
-@receiver(pre_save, sender=Site)
-def request_update_document_visibility(sender, instance, **kwargs):
-    if instance.id is None:
-        # New site, don't do anything
-        return
-
-    try:
-        original_site = Site.objects.get(id=instance.id)
-    except ObjectDoesNotExist:
-        # New site or deleted, don't do anything
-        return
-
-    if original_site.visibility != instance.visibility:
-        update_document_visibility.apply_async(
-            (instance.id, instance.visibility), link_error=link_error_handler.s()
-        )
 
 
 @shared_task
@@ -58,15 +39,6 @@ def update_document_visibility(instance_id, instance_visibility, **kwargs):
     # updating story visibility
     for story in stories_set:
         update_story_visibility(story, instance_visibility)
-
-
-# If a site is deleted, delete all docs from index related to site
-@receiver(post_delete, sender=Site)
-def request_delete_related_docs(sender, instance, **kwargs):
-    if Site.objects.filter(id=instance.id).exists():
-        delete_related_docs.apply_async(
-            (instance.id,), link_error=link_error_handler.s()
-        )
 
 
 @shared_task
