+++ firstvoices/backend/tests/conftest.py
@@ -40,35 +40,35 @@ def mock_get_image_dimensions(mocker):
 @pytest.fixture(autouse=True, scope="session")
 def mock_search_indexing():
     with patch(
-        "backend.search.indices.dictionary_entry_document.update_dictionary_entry_index"
+        "backend.search.signals.dictionary_entry_signals.update_dictionary_entry_index"
     ) as mocked_update_dictionary_entry_index, patch(
-        "backend.search.indices.dictionary_entry_document.delete_from_index"
+        "backend.search.signals.dictionary_entry_signals.delete_from_index"
     ) as mocked_delete_dictionary_entry_index, patch(
-        "backend.search.indices.dictionary_entry_document.update_translation"
+        "backend.search.signals.dictionary_entry_signals.update_translation"
     ) as mocked_update_translation, patch(
-        "backend.search.indices.dictionary_entry_document.update_notes"
+        "backend.search.signals.dictionary_entry_signals.update_notes"
     ) as mocked_update_notes, patch(
-        "backend.search.indices.dictionary_entry_document.update_acknowledgements"
+        "backend.search.signals.dictionary_entry_signals.update_acknowledgements"
     ) as mocked_update_acknowledgements, patch(
-        "backend.search.indices.dictionary_entry_document.update_categories"
+        "backend.search.signals.dictionary_entry_signals.update_categories"
     ) as mocked_update_categories, patch(
-        "backend.search.indices.dictionary_entry_document.update_categories_m2m"
+        "backend.search.signals.dictionary_entry_signals.update_categories_m2m"
     ) as mocked_update_categories_m2m, patch(
-        "backend.search.indices.song_document.update_song_index"
+        "backend.search.signals.song_signals.update_song_index"
     ) as mocked_update_song_index, patch(
-        "backend.search.indices.song_document.delete_from_index"
+        "backend.search.signals.song_signals.delete_from_index"
     ) as mocked_delete_song_index, patch(
-        "backend.search.indices.song_document.update_lyrics"
+        "backend.search.signals.song_signals.update_lyrics"
     ) as mocked_update_lyrics, patch(
-        "backend.search.indices.story_document.update_story_index"
+        "backend.search.signals.story_signals.update_story_index"
     ) as mocked_update_story_index, patch(
-        "backend.search.indices.story_document.delete_from_index"
+        "backend.search.signals.story_signals.delete_from_index"
     ) as mocked_delete_story_index, patch(
-        "backend.search.indices.story_document.update_pages"
+        "backend.search.signals.story_signals.update_pages"
     ) as mocked_request_update_pages, patch(
-        "backend.search.utils.site_signals.update_document_visibility"
+        "backend.search.signals.site_signals.update_document_visibility"
     ) as mocked_update_document_visibility, patch(
-        "backend.search.utils.site_signals.delete_related_docs"
+        "backend.search.signals.site_signals.delete_related_docs"
     ) as mocked_delete_related_docs:
         mocked_update_dictionary_entry_index.return_value = None
         mocked_delete_dictionary_entry_index.return_value = None
